--- src/util/install.js
@@ -1,18 +1,0 @@
-const fs = require('fs')
-
-const { getExtractionPath } = require('../util/utils')
-const { yvmPath } = require('../util/path')
-
-const installVersion = require('../commands/install')
-
-const ensureVersionInstalled = (version, rootPath = yvmPath) => {
-    const yarnBinDir = getExtractionPath(version, rootPath)
-    if (fs.existsSync(yarnBinDir)) {
-        return Promise.resolve()
-    }
-    return installVersion(version, rootPath)
-}
-
-module.exports = {
-    ensureVersionInstalled,
-}
