+++ test/commands/install.test.js
@@ -1,6 +1,6 @@
 const fs = require('fs-extra')
 
-const { install, installLatest } = require('../../src/commands/install')
+const { installVersion, installLatest } = require('../../src/commands/install')
 const {
     getVersionsFromTags,
     getExtractionPath,
@@ -20,7 +20,7 @@ describe('yvm install', () => {
 
     it('Installs a valid yarn version', () => {
         const version = '1.7.0'
-        return install(version, rootPath).then(() => {
+        return installVersion(version, rootPath).then(() => {
             expect(
                 fs.statSync(getExtractionPath(version, rootPath)),
             ).toBeTruthy()
@@ -30,21 +30,18 @@ describe('yvm install', () => {
     it('Installs two versions of Yarn', () => {
         const v1 = '1.7.0'
         const v2 = '1.6.0'
-        return Promise.all([install(v1, rootPath), install(v2, rootPath)]).then(
-            () => {
-                expect(
-                    fs.statSync(getExtractionPath(v1, rootPath)),
-                ).toBeTruthy()
-                expect(
-                    fs.statSync(getExtractionPath(v2, rootPath)),
-                ).toBeTruthy()
-            },
-        )
+        return Promise.all([
+            installVersion(v1, rootPath),
+            installVersion(v2, rootPath),
+        ]).then(() => {
+            expect(fs.statSync(getExtractionPath(v1, rootPath))).toBeTruthy()
+            expect(fs.statSync(getExtractionPath(v2, rootPath))).toBeTruthy()
+        })
     })
 
     it('Installs doesnt install an invalid version of Yarn', () => {
         const version = '0.0.0'
-        return install(version, rootPath).catch(() => {
+        return installVersion(version, rootPath).catch(() => {
             expect(() =>
                 fs.statSync(getExtractionPath(version, rootPath)),
             ).toThrow()
